A Mini and a Tiny arcade cabinet :)
Here are two of my first 3D sculptures for the Oculus Home!

The Mini arcade cabinet was designed to scale around a 7 inch Raspberry Pi screen and regular arcade controls, although I think the sizing is a little off after converting the file a few times :/

Feel free to download and enjoy!